@@ -6,6 +6,6 @@ namespace VacationRental.Api.Models
     public class CalendarDateViewModel
     {
         public DateTime Date { get; set; }
-        public List<CalendarBookingViewModel> Bookings { get; set; }
+        public List<CalendarBookingViewModel>? Bookings { get; set; }
     }
 }
