@@ -5,6 +5,6 @@ namespace VacationRental.Api.Models
     public class CalendarViewModel
     {
         public int RentalId { get; set; }
-        public List<CalendarDateViewModel> Dates { get; set; }
+        public List<CalendarDateViewModel>? Dates { get; set; }
     }
 }
